fix some issues of shared heap

+++ core/iwasm/interpreter/wasm_runtime.h
@@ -97,7 +97,7 @@ typedef struct WASMSharedHeap {
     struct WASMSharedHeap *next;
     void *heap_handle;
     uint8 *base_addr;
-    uint32 size;
+    uint64 size;
     uint64 start_off_mem64;
     uint64 start_off_mem32;
 } WASMSharedHeap;
